// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Wed, 12 Jul 2023 07:08:16 GMT
 */

/**
 * @enum
 * @readonly
 * @description RefWorlds Action Names
 */
export enum RefWorldsActionName {
    Cancel = 'cancel',
    Clean = 'clean',
    Clearconfig = 'clearconfig',
    Exec = 'exec',
    Propose = 'propose',
    Publresult = 'publresult',
    Refund = 'refund',
    Rmvexpired = 'rmvexpired',
    Stakeobsv = 'stakeobsv',
    Updateconfig = 'updateconfig',
    Updatestatus = 'updatestatus',
    Vote = 'vote',
};